+++ public/js/app.js
@@ -2291,7 +2291,7 @@ __webpack_require__.r(__webpack_exports__);
       }
     },
     toggleHighlightComponent: function toggleHighlightComponent(_index) {
-      var _verse = document.getElementById('verse-' + _index);
+      var _verse = document.getElementById('verse-' + (_index + 1));
 
       var _verseClasses = _verse.classList;
 
@@ -2387,6 +2387,7 @@ __webpack_require__.r(__webpack_exports__);
   methods: {
     getSearch: function getSearch() {
       var words = this.query.split(' ');
+      var versePattern = /abc/;
       var q = '';
 
       for (var i = 0; i < words.length; i++) {
@@ -2404,6 +2405,44 @@ __webpack_require__.r(__webpack_exports__);
   mounted: function mounted() {
     console.log('Component mounted.');
   }
+});
+
+/***/ }),
+
+/***/ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=script&lang=js&":
+/*!***************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/babel-loader/lib??ref--4-0!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=script&lang=js& ***!
+  \***************************************************************************************************************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+//
+/* harmony default export */ __webpack_exports__["default"] = ({
+  mounted: function mounted() {
+    console.log(this.results);
+  },
+  props: ['results']
 });
 
 /***/ }),
@@ -38456,7 +38495,7 @@ var render = function() {
                     "p",
                     {
                       staticClass: "verse",
-                      attrs: { id: "verse-" + index },
+                      attrs: { id: "verse-" + (index + 1) },
                       on: {
                         click: function($event) {
                           return _vm.toggleUnderline(text.id)
@@ -38645,6 +38684,54 @@ var render = function() {
           )
         ]
       )
+    ])
+  ])
+}
+var staticRenderFns = []
+render._withStripped = true
+
+
+
+/***/ }),
+
+/***/ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=template&id=318ad535&":
+/*!*******************************************************************************************************************************************************************************************************************************!*\
+  !*** ./node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!./node_modules/vue-loader/lib??vue-loader-options!./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=template&id=318ad535& ***!
+  \*******************************************************************************************************************************************************************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "render", function() { return render; });
+/* harmony export (binding) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return staticRenderFns; });
+var render = function() {
+  var _vm = this
+  var _h = _vm.$createElement
+  var _c = _vm._self._c || _h
+  return _c("div", { staticClass: "container" }, [
+    _c("div", { staticClass: "row justify-content-center" }, [
+      _c("div", { staticClass: "col-md-8" }, [
+        _c("div", { staticClass: "card" }, [
+          _c("div", { staticClass: "card-header" }, [_vm._v("Search Results")]),
+          _vm._v(" "),
+          _c(
+            "div",
+            { staticClass: "card-body" },
+            [
+              _vm._v(
+                "\n                    " +
+                  _vm._s(_vm.results) +
+                  " search results found.\n                    "
+              ),
+              _vm._l(_vm.results, function(result) {
+                return _c("article", { key: result.id })
+              })
+            ],
+            2
+          )
+        ])
+      ])
     ])
   ])
 }
@@ -50830,6 +50917,7 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony import */ var _components_Bible_PageComponent_vue__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./components/Bible/PageComponent.vue */ "./resources/js/components/Bible/PageComponent.vue");
 /* harmony import */ var _components_Bible_SavedHighlightsComponent_vue__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./components/Bible/SavedHighlightsComponent.vue */ "./resources/js/components/Bible/SavedHighlightsComponent.vue");
 /* harmony import */ var _components_Bible_SearchComponent_vue__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./components/Bible/SearchComponent.vue */ "./resources/js/components/Bible/SearchComponent.vue");
+/* harmony import */ var _components_Bible_SearchResultsComponent_vue__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./components/Bible/SearchResultsComponent.vue */ "./resources/js/components/Bible/SearchResultsComponent.vue");
 /**
  * First we will load all of this project's JavaScript dependencies which
  * includes Vue and other libraries. It is a great starting point when
@@ -50855,6 +50943,7 @@ window.Vue = __webpack_require__(/*! vue */ "./node_modules/vue/dist/vue.common.
 // Vue.component('leaf-component', require('./components/Bible/LeafComponent.vue').default);
 // Vue.component('login-component', require('./components/Form/LoginComponent.vue').default);
 // Import local components.
+
 
 
 
@@ -50886,7 +50975,8 @@ var app = new Vue({
     'index-component': _components_Bible_IndexComponent_vue__WEBPACK_IMPORTED_MODULE_0__["default"],
     'page-component': _components_Bible_PageComponent_vue__WEBPACK_IMPORTED_MODULE_1__["default"],
     'saved-highlights-component': _components_Bible_SavedHighlightsComponent_vue__WEBPACK_IMPORTED_MODULE_2__["default"],
-    'search-component': _components_Bible_SearchComponent_vue__WEBPACK_IMPORTED_MODULE_3__["default"]
+    'search-component': _components_Bible_SearchComponent_vue__WEBPACK_IMPORTED_MODULE_3__["default"],
+    'search-results-component': _components_Bible_SearchResultsComponent_vue__WEBPACK_IMPORTED_MODULE_4__["default"]
   },
   data: {
     books: null,
@@ -51387,6 +51477,75 @@ __webpack_require__.r(__webpack_exports__);
 /* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_SearchComponent_vue_vue_type_template_id_daac6d42___WEBPACK_IMPORTED_MODULE_0__["render"]; });
 
 /* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_SearchComponent_vue_vue_type_template_id_daac6d42___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });
+
+
+
+/***/ }),
+
+/***/ "./resources/js/components/Bible/SearchResultsComponent.vue":
+/*!******************************************************************!*\
+  !*** ./resources/js/components/Bible/SearchResultsComponent.vue ***!
+  \******************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _SearchResultsComponent_vue_vue_type_template_id_318ad535___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./SearchResultsComponent.vue?vue&type=template&id=318ad535& */ "./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=template&id=318ad535&");
+/* harmony import */ var _SearchResultsComponent_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./SearchResultsComponent.vue?vue&type=script&lang=js& */ "./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=script&lang=js&");
+/* empty/unused harmony star reexport *//* harmony import */ var _node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ../../../../node_modules/vue-loader/lib/runtime/componentNormalizer.js */ "./node_modules/vue-loader/lib/runtime/componentNormalizer.js");
+
+
+
+
+
+/* normalize component */
+
+var component = Object(_node_modules_vue_loader_lib_runtime_componentNormalizer_js__WEBPACK_IMPORTED_MODULE_2__["default"])(
+  _SearchResultsComponent_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_1__["default"],
+  _SearchResultsComponent_vue_vue_type_template_id_318ad535___WEBPACK_IMPORTED_MODULE_0__["render"],
+  _SearchResultsComponent_vue_vue_type_template_id_318ad535___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"],
+  false,
+  null,
+  null,
+  null
+  
+)
+
+/* hot reload */
+if (false) { var api; }
+component.options.__file = "resources/js/components/Bible/SearchResultsComponent.vue"
+/* harmony default export */ __webpack_exports__["default"] = (component.exports);
+
+/***/ }),
+
+/***/ "./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=script&lang=js&":
+/*!*******************************************************************************************!*\
+  !*** ./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=script&lang=js& ***!
+  \*******************************************************************************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_SearchResultsComponent_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/babel-loader/lib??ref--4-0!../../../../node_modules/vue-loader/lib??vue-loader-options!./SearchResultsComponent.vue?vue&type=script&lang=js& */ "./node_modules/babel-loader/lib/index.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=script&lang=js&");
+/* empty/unused harmony star reexport */ /* harmony default export */ __webpack_exports__["default"] = (_node_modules_babel_loader_lib_index_js_ref_4_0_node_modules_vue_loader_lib_index_js_vue_loader_options_SearchResultsComponent_vue_vue_type_script_lang_js___WEBPACK_IMPORTED_MODULE_0__["default"]); 
+
+/***/ }),
+
+/***/ "./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=template&id=318ad535&":
+/*!*************************************************************************************************!*\
+  !*** ./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=template&id=318ad535& ***!
+  \*************************************************************************************************/
+/*! exports provided: render, staticRenderFns */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+__webpack_require__.r(__webpack_exports__);
+/* harmony import */ var _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_SearchResultsComponent_vue_vue_type_template_id_318ad535___WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! -!../../../../node_modules/vue-loader/lib/loaders/templateLoader.js??vue-loader-options!../../../../node_modules/vue-loader/lib??vue-loader-options!./SearchResultsComponent.vue?vue&type=template&id=318ad535& */ "./node_modules/vue-loader/lib/loaders/templateLoader.js?!./node_modules/vue-loader/lib/index.js?!./resources/js/components/Bible/SearchResultsComponent.vue?vue&type=template&id=318ad535&");
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "render", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_SearchResultsComponent_vue_vue_type_template_id_318ad535___WEBPACK_IMPORTED_MODULE_0__["render"]; });
+
+/* harmony reexport (safe) */ __webpack_require__.d(__webpack_exports__, "staticRenderFns", function() { return _node_modules_vue_loader_lib_loaders_templateLoader_js_vue_loader_options_node_modules_vue_loader_lib_index_js_vue_loader_options_SearchResultsComponent_vue_vue_type_template_id_318ad535___WEBPACK_IMPORTED_MODULE_0__["staticRenderFns"]; });